backend/requirements.txt
flask
flask-cors
pytesseract
pdfplumber
pdf2image
pillow
torch
transformers
datasets
scikit-learn
pandas